break platform cfg code into utils submodule
